codegen+typechecker: Add support for attributes and 'extern names'
These attributes can be applied to function and struct/classes (for
now), and allow the user to specify the generated name of the function.
This is exceptionally helpful for interfacing with C++ libraries that
have annoying names that you don't want to directly deal with :^)

This commits also adds the 'deprecated()' attribute, which makes the
compiler yell at you for using the function (but allows typechecking to
continue).