parser: Allow trailing closure arguments to occur after the call's ()
See `samples/closures/trailing_closure_parameter.jakt` for an example.

As shown by the tests, these are not parsed in the expressions of
if/for/guard/match, since that leads to ambiguous syntax and also just
looks bad.

This is a very limited first pass. A single block is parsed, and treated
as a closure. This means that closure parameters, captures, and the
return type (including the throwingness) cannot be defined. Ideally the
return type would be inferred based on the closure's contents and the
called function's parameter type.