typechecker: Don't fallback to the argument name, if label is present
...while validating the argument label when label is required.
Achieve this by using the recently added get_argument_name().
This has the following effects:

1. We can no longer override label name if the argument name matches
the parameter name. This was fixed in 5c0b593 but only in situations
where label was explicitly not required (ex. parameter marked as `anon`)

2. Previously, when printing an error when the label didn't match
the parameter name and the argument name was retrievable and didn't
match the parameter name either, we would print an error highlighting
the *argument* name instead of the wrong label. Now in all situations
we highlight the wrong label. If the label is not present the error
has a new form. Instead of:
"Wrong parameter name in argument label got 'arg_name', expected
'param_name'"
it is:
"Missing argument label (expected 'param_name:')"