typechecker: Allow field references in methods
Allow creating local references to struct fields in methods.

It turned out the way we were checking scope lifetimes was a bit
backwards. We use this method every to guard against the inverse (the
child scope outliving the parent) but the method wasn't implemented that
way. To be honest I still don't understand a 100% why some tests were
passing in the old implementation.

We I also had to invert the exit condition to account for the fact that
comptime scope parent chains are disjunct from runtime scope parent
chains, but at least we now know that the json parser sample is testing
for that case :wink:.

Fixing this also surfaced that the `stores_arguments` attribute test
wasn't actually testing the feature but was passing due to the bug. I
had to adjust it to actually test the failure case for it to continue to
pass.

Progress on #1379.