selfhost: Add `is Some` operator
Make it possible to match on optionals using `is Some` conditions. As if
they were enums we can now say
```
if a is Some(x) {
  println(x)
}
```

or

```
if a is Some {
  println("Has Value")
}
```

This should hopefully make a bunch of `.has_value()` calls followed by
forced unwraps everywhere a thing of the past.

Like for "normal" `is MyEnum(x)` conditions, it will currently **not**
possible to nest the matcher (`is Some(Some(x))`).