CMake: Enable all warnings, then manually disable those we trigger
When building Serenity with ENABLE_JAKT, it builds Jakt using this cmake
file, BUT it already has the `-Wall -Wextra -Werror` options set. This
was causing it to fail to compile Jakt: Some warnings were not enabled
by default, so building Jakt directly would work, but they are enabled
by these extra options, so would fail as part of building Serenity.

By defining the options here, we make sure that we can't accidentally
break the Serenity build without also breaking the Jakt build itself.

Of course, ideally we would also generate code that doesn't set off
warnings, but that's trickier.