typechecker: Defer struct default field typecheck
This overcomes #1429 by inserting temporary garbage expressions into the
default value of the parameters in the generated struct constructor. To
make this work, I had to add a type id to the `Garbage` checked
expression, so that constructor calls that are checked before the
default field being filled can be properly resolved.