selfhost: Make unqualified scope lookup more accurate
Previously we also allowed children of scopes accessible with
unqualified lookup to be used unqualified, which made a complete mess
when the closest scope didn't have the entity that was looked up.
For instance, prelude::Reflect::Span() was chosen when the input was
'AK::Span()', due to `Reflect` being accessible from the lookup site.

This commit fixes this behaviour, and also ensures that externally
imported code doesn't get confused with existing namespaces and types.