codegen: Simplify control flow forwarding rules for match IIFEs
ControlFlowState was handling control flow forwarding in an error-prone
and unintuitive manner. It also was being used to apply control flow
forwarding from the scope of the match IIFE that is being called,
instead of the state of the outer scope, which is where the calling and
forwarding takes place.

The only tracking that is needed to apply control flow forwarding is the
following two things about the parent scope:
    1. Is the current scope the outer loop where continue/break is
       handled or should it forward the special LoopContinue/LoopBreak
       signaling types?
    2. Is the current scope the function scope to return from or should
       it forward the special signaling return from
       `ExplicitValueOrControlFlow`?