codegen: Use the correct name to search for common enum members
In variant pattern arguments, the name was set to "value" on
non-StructLike enum variants, which means that the common enum member
that the binding might have referred to is never found, and thus the
wrong code is emmited: `.as.#Variant.value`. The correct code to emit
when we have an enum member is `.common.init_common.#member_name`.