lexer: Use common enum member for Token
This will improve compile times since with common enum members the
code generated to access them is simpler. They also are guaranteed to
exist in any variant, so the spans can be just accessed like a regular
member and no matching is required.

Since Token also has a mix of variant kinds (untyped, structlike) this
serves as a way to easily see where codegen wastes C++ tokens in saying
nothing.