codegen: Remove uninit check in enums
The safety mechanism that was introduced to know if an enum is uninit
isn't needed, since all the paths that may have an uninit value are
contained in the constructor functions, and the external code uses the
constructor functions to get new values.

Thanks to this, most of the custom copy/move/assignment constructors can
be removed if there is no union initialization / destruction to do.