codegen: Make common enum fields be struct fields directly
This avoids union initialization in the common values, which was the old
solution to make the whole enum default initializable. Now we can leave
the oddities to just the VariantData part and have the common data
initialized via normal C++ constructors.

This manages to shave off almost 10s of codegen time in the C++ end.