parser+runtime: Add `appendff` under `StringBuilder`
This way we can remove a lot of duplicated effort where we format into a
String to then copy the string directly to the StringBuffer.

The parser had a little choke, where it would error in the rparen after
varargs: `..)`. It thought it had just passed over a comma, which was
the only possible case until now. `runtime/prelude.jakt` has `appendff`
which is the test for it passing.