codegen: Move forward declarations to their respective headers
The "predecl" declarations that resolve cyclic dependencies (e.g
tests/codegen/ns_circular_dependency.jakt) are now emitted on each
header, so that we don't include all the baggage from other modules when
we don't need it. It also avoids polluting the preprocessor's namespace.

Codegen now tracks the dependencies it needs for the C++ file generation
as the module header only includes what's necessary for the type and
function definitions, not for the implementations. After tracking,
codegen includes in the C++ implementation only the headers that weren't
already included by the module header. Previously, this was partly done
(without the set difference) when generating
`<module>_specializations.cpp`, but not in `<module>.cpp` .

This removes the need for a __unified_forward.h header, which right now
only serves the purpose of defining SetConsoleOutputCP for the file with
the Jakt::main function.