codegen: Register modules through used functions too
When generating C++ implementations, the functions from a module may be
used without explicitly stating any of the types of said module. Modules
can also be comprised only by functions. Now codegen adds the function's
module whenever it calls a function to make sure that the header for
that function is added.