Jakttest: Add an option to keep temp directories
When debugging the code generator, it keeps getting very annoying
that Jakttest cleans the temporary directory. It is much more useful
to tell it to keep the artifacts in place, instead of having to
replicate the flags that Jakttest used in a separate command.

This flag can be used in conjunction with `--temp-dir` and a single test
file so that the test artifacts for said file can be observed after the
test has run. Using it under bulk testing is advised against, since
Jakttest may have cleaned the directory to run a different test.