selfhost: Do not conflate module namespace names with regular names
A module namespace is not inherently named, that's an implementation
detail that should not be exposed to the user.
Previously `import foo { foo }` would import the module `foo`'s main
namespace, which was effectively the same as `import foo`; this is not a
desirable behaviour as any namespace called `foo` within that module
would effectively be shadowed by this alias.