codegen: Include toplevel `Jakt::` in namespace qualifiers
Namespace qualifiers for Jakt types (not prelude ones) now have the
`Jakt::` namespace prepended to them.

Fixes a bug where if a class A is defined in module A, then: Jakt would
use `A::A` in the method definition of a class, and C++ would complain
saying that `A::A` refers to the constructor, since the method
definition is inside the class.