selfhost: Unbork module loading
Previously a module's "sanitised" name only depended on its input name
(from the source code), this was correct when the only way to address
imports was to start from the root, but the addition of the standard lib
and relative imports means that the input name is no longer a unique
name for the module.
This commit makes the sanitised name depend on the real path of the
module file:
- If the file is in the standard lib, its name is `jakt/path/in/stdlib`
- Otherwise, its name is the relative path from the project root to the
  file, without the file extension.

This is compatible with the existing system when normal imports are
used, but relative imports will no longer use the relative name (which
was also a problem with module caching).