codegen: Hint when parentheses are needed in codegen_expression
codegen_expression now will only emit parentheses if the expression that
is emitted must be self contained syntactically and it is not by
default.

This removes most wasteful uses of parentheses that make conditions from
`if` and `while` hard to parse visually due to a large amount of opening
parentheses.