codegen: Do not use IIFE in matches that yield void
We don't need to have any value at the end of the match so there's no
point of using an IIFE. This allows us to use regular `break`,
`continue` and `return` inside them without having to resort to
ExplicitValueOrControlFlow (EVOCF), which also reduces the amount of
template instantiatons that the C++ compiler has to do.

Since the code that bootstrap generates for stage1 still generates EVOCF
on void matches, we cannot remove the `ExplicitValue<void>` handling in
runtime/Jakt/ControlFlow.h without a rebootstrap, which is not yet
needed.