typechecker+codegen: Emit new bodies for bindings with different types
This fixes cases where the body is not duplicated when it should,
specially with instantiations that are inferred inside the body for
functions that support both instantiations and overloading. C++ can deal
with overloading, but we add the template parameters ourselves, which
makes the typing wrong for at least one pattern.

Introduces `BindingKey`, `MatchBuilder and `BindingKeyBuilder`, which
together help keep match body code duplication at a low rate while being
as least intrusive as possible to the logic for each match typecheck:
`typecheck_match` has three big `for` loops with bodies that walk the
match cases in very different ways, and the deduplication system is
orthogonal to all of them. Most of it is implemented between
`MatchBuilder` and `BindingKeyBuilder`.

These builder states avoid creating scopes and duplicate typechecker
work, resulting in ~14% performance increase and 20MB less peak memory
usage when typechecking selfhost.