selfhost: Blast find_or_add_type_id() off profiles
This speeds up find_or_add_type_id() using one major observation and
two observations:
  - A type does not change significantly once it's created
  - The type can have a more lax hash that ignores large parts of the
    type
  - There aren't that many types with the same lax hash actually

In this commit, the type list in a module gains a skiplist where each
"lane" is reserved for a specific hash; this list grows linearly with
the number of unique hashes, as each entry tracks its own next-in-list
object.

Overall, this is a ~40% performance increase when typechecking the
compiler, and a ~11% increase when typechecking all the tests in the
repository.