cmake: Precompile `runtime/lib.h`
It is a big header included in all C++ files generated by Jakt, so when
profiling a C++ build it takes nearly 40% of the time for frontend
analysis.

Although it would be nice to configure it in `runtime/lib.h`, I could
not get it to work with the same line, so it'd be appreciated.

This is the low hanging fruit of more steps we could take towards making
the build share more code. Automatically deciding on where to put PCH
files sounds interesting but introduces a burden on configuring via
cmake. Another solution I can think of is C++ modules, although that is
going to need more iteration.